Handle non-standard trailing characters in the version list
There are a few "bad" version numbers in Module::CoreList. Trailing
spaces emit warnings- but at the moment those trailing spaces should be
left in, as they indicate a likely CVS source, which a future hacker
might want to convert into "v"-strings.

There is also one module that was released with a trailing letter "a",
and version.pm only wants digits, "." and ".", so this patch converts a
trailing letter to underscore plus its ASCII code.